% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fmriqa.R
\name{combine_res_glob}
\alias{combine_res_glob}
\title{Combine fmriqa csv result files into a dataframe}
\usage{
combine_res_glob(pattern)
}
\arguments{
\item{pattern}{glob pattern to match the csv result files (eg "*.csv")}
}
\value{
a dataframe of results
}
\description{
Combine fmriqa csv result files into a dataframe
}
