% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmriqa.R
\docType{package}
\name{fmriqa-package}
\alias{fmriqa}
\alias{fmriqa-package}
\title{fmriqa: fMRI quality assurance routines}
\description{
For a full list of functions:
`help(package=fmriqa, help_type="html")`
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/martin3141/fmriqa/issues}
}

}
\author{
\strong{Maintainer}: Martin Wilson \email{martin@pipegrep.co.uk}

}
