% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fmriqa.R
\name{run_fmriqa_glob}
\alias{run_fmriqa_glob}
\title{Run fMRI quality assurance procedure on a set of NIfTI data files}
\usage{
run_fmriqa_glob(pattern, ...)
}
\arguments{
\item{pattern}{glob expresion to match analysis files}

\item{...}{options to pass to run_fmriqa function}
}
\description{
Run fMRI quality assurance procedure on a set of NIfTI data files
}
