% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qa.R
\name{run_qa}
\alias{run_qa}
\title{Run fMRI quality assurance procedure on a NIfTI data file}
\usage{
run_qa(data_file = NULL, roi_width = 21, slice_num = NULL, skip = 2,
  tr = NULL, spike_detect = FALSE, x_pos = NULL, y_pos = NULL,
  plot_title = NULL, last_vol = NULL, gen_png = TRUE,
  gen_res_csv = TRUE, gen_pdf = FALSE, gen_spec_csv = FALSE,
  res_fname = NULL, png_fname = NULL, pdf_fname = NULL,
  spec_fname = NULL)
}
\arguments{
\item{data_file}{a}

\item{roi_width}{b}

\item{slice_num}{c}

\item{skip}{d}

\item{tr}{e}

\item{spike_detect}{f}

\item{x_pos}{g}

\item{y_pos}{h}

\item{plot_title}{i}

\item{last_vol}{j}

\item{gen_png}{k}

\item{gen_res_csv}{l}

\item{gen_pdf}{m}

\item{gen_spec_csv}{n}

\item{res_fname}{o}

\item{png_fname}{p}

\item{pdf_fname}{q}

\item{spec_fname}{r}
}
\value{
dataframe of QA metrics
}
\description{
Run fMRI quality assurance procedure on a NIfTI data file
}
