% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmriqa.R
\docType{package}
\name{fmriqa-package}
\alias{fmriqa}
\alias{fmriqa-package}
\title{fmriqa: fMRI quality assurance routines}
\description{
The fmriqa package provides an implemenation of the fMRI QA analysis protocol
detailed in : Friedman L and Glover GH. Report on a multicenter fMRI quality
assurance protocol. J Magn Reson Imaging. 2006 Jun;23(6):827-39.
}
\details{
For a full list of functions:
`help(package=fmriqa, help_type="html")`
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/martin3141/fmriqa/issues}
}

}
\author{
\strong{Maintainer}: Martin Wilson \email{martin@pipegrep.co.uk}

}
