% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmriqa.R
\docType{package}
\name{fmriqa-package}
\alias{fmriqa}
\alias{fmriqa-package}
\title{fmriqa: fMRI quality assurance routines}
\description{
The fmriqa package provides an implementation of the fMRI quality assurance
analysis protocol detailed by Friedman and Glover (2006)
<doi:10.1002/jmri.20583>.
}
\details{
#' To learn more about fmriqa, start with the vignettes:
`browseVignettes(package = "fmriqa")`

For a full list of functions:
`help(package=fmriqa, help_type="html")`
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/martin3141/fmriqa/issues}
}

}
\author{
\strong{Maintainer}: Martin Wilson \email{martin@pipegrep.co.uk}

}
